src/main/java/leetcode/Solution00862.java
/**
 * 862. Shortest Subarray with Sum at Least K
 * <a href="https://leetcode.com/problems/shortest-subarray-with-sum-at-least-k">Web link</a> |
 * Hard
 */

package leetcode;

import java.util.ArrayDeque;
import java.util.Deque;

public class Solution00862 {

    public int shortestSubarray(int[] nums, int k) {
        int n = nums.length;
        long[] prefix = new long[n + 1];

        // Step 1: Compute prefix sums
        for (int i = 0; i < n; i++) {
            prefix[i + 1] = prefix[i] + nums[i];
        }

        Deque<Integer> deque = new ArrayDeque<>();
        int minLength = Integer.MAX_VALUE;

        // Step 2: Process prefix sums
        for (int i = 0; i <= n; i++) {
            while (!deque.isEmpty() && prefix[i] - prefix[deque.peekFirst()] >= k) {
                minLength = Math.min(minLength, i - deque.pollFirst());
            }

            while (!deque.isEmpty() && prefix[i] <= prefix[deque.peekLast()]) {
                deque.pollLast();
            }

            deque.addLast(i);
        }

        return minLength == Integer.MAX_VALUE ? -1 : minLength;
    }

}